+++ mcm4csharp/v1/Data/Content/ConversationContent.cs
@@ -4,7 +4,7 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Content {
 	public struct ConversationContent {
 		[JsonPropertyName("recipient_ids")]
-		public uint[] RecipientIds { get; set; }
+		public ulong[] RecipientIds { get; set; }
 
 		[JsonPropertyName("title")]
 		public string Title { get; set; }
