+++ mcm4csharp/v1/Data/Conversations/Conversation.cs
@@ -4,31 +4,31 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Conversations {
 	public struct Conversation {
 		[JsonPropertyName ("conversation_id")]
-		public uint ConversationId { get; set; }
+		public ulong ConversationId { get; set; }
 
 		[JsonPropertyName ("title")]
 		public string Title { get; set; }
 
 		[JsonPropertyName ("creation_date")]
-		public uint CreationDate { get; set; }
+		public ulong CreationDate { get; set; }
 
 		[JsonPropertyName ("creator_id")]
-		public uint CreatorId { get; set; }
+		public ulong CreatorId { get; set; }
 
 		[JsonPropertyName ("last_message_date")]
-		public uint LastMessageDate { get; set; }
+		public ulong LastMessageDate { get; set; }
 
 		[JsonPropertyName ("last_read_date")]
-		public uint LastReadDate { get; set; }
+		public ulong LastReadDate { get; set; }
 
 		[JsonPropertyName ("open")]
 		public bool Open { get; set; }
 
 		[JsonPropertyName ("reply_count")]
-		public uint ReplyCount { get; set; }
+		public ulong ReplyCount { get; set; }
 
 		[JsonPropertyName ("recipient_ids")]
-		public uint [] RecipientIds { get; set; }
+		public ulong [] RecipientIds { get; set; }
 	}
 }
