+++ mcm4csharp/v1/Data/Members/Member.cs
@@ -4,16 +4,16 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Members {
 	public struct Member {
 		[JsonPropertyName ("member_id")]
-		public uint MemberId { get; set; }
+		public ulong MemberId { get; set; }
 
 		[JsonPropertyName ("username")]
 		public string Username { get; set; }
 
 		[JsonPropertyName ("join_date")]
-		public uint JoinDate { get; set; }
+		public ulong JoinDate { get; set; }
 
 		[JsonPropertyName ("last_activity_date")]
-		public uint? LastActivityDate { get; set; }
+		public ulong? LastActivityDate { get; set; }
 
 		[JsonPropertyName ("banned")]
 		public bool Banned { get; set; }
@@ -25,22 +25,22 @@ namespace mcm4csharp.v1.Data.Members {
 		public bool Disabled { get; set; }
 
 		[JsonPropertyName ("post_count")]
-		public uint PostCount { get; set; }
+		public ulong PostCount { get; set; }
 
 		[JsonPropertyName ("resource_count")]
-		public uint ResourceCount { get; set; }
+		public ulong ResourceCount { get; set; }
 
 		[JsonPropertyName ("purchase_count")]
-		public uint PurchaseCount { get; set; }
+		public ulong PurchaseCount { get; set; }
 
 		[JsonPropertyName ("feedback_positive")]
-		public uint PositiveFeedback { get; set; }
+		public ulong PositiveFeedback { get; set; }
 
 		[JsonPropertyName ("feedback_neutral")]
-		public uint NeutralFeedback { get; set; }
+		public ulong NeutralFeedback { get; set; }
 
 		[JsonPropertyName ("feedback_negative")]
-		public uint NegativeFeedback { get; set; }
+		public ulong NegativeFeedback { get; set; }
 	}
 }
