+++ mcm4csharp/v1/Data/Resources/License.cs
@@ -4,10 +4,10 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Resources {
 	public struct License {
 		[JsonPropertyName ("license_id")]
-		public uint LicenseId { get; set; }
+		public ulong LicenseId { get; set; }
 
 		[JsonPropertyName ("purchaser_id")]
-		public uint PurchaserId { get; set; }
+		public ulong PurchaserId { get; set; }
 
 		// some bizarre choice for "Get resource license by member", add optional thing here.
 		[JsonPropertyName("purchaser_name")]
@@ -20,13 +20,13 @@ namespace mcm4csharp.v1.Data.Resources {
 		public bool Active { get; set; }
 
 		[JsonPropertyName ("start_date")]
-		public uint StartDate { get; set; }
+		public ulong StartDate { get; set; }
 
 		[JsonPropertyName ("end_date")]
-		public uint EndDate { get; set; }
+		public ulong EndDate { get; set; }
 
 		[JsonPropertyName ("previous_end_date")]
-		public uint PreviousEndDate { get; set; }
+		public ulong PreviousEndDate { get; set; }
 	}
 }
