+++ mcm4csharp/v1/Data/Resources/Review.cs
@@ -4,16 +4,16 @@ using System.Text.Json.Serialization;
 namespace mcm4csharp.v1.Data.Resources {
 	public struct Review {
 		[JsonPropertyName ("review_id")]
-		public uint ReviewId { get; set; }
+		public ulong ReviewId { get; set; }
 
 		[JsonPropertyName ("reviewer_id")]
-		public uint ReviewerId { get; set; }
+		public ulong ReviewerId { get; set; }
 
 		[JsonPropertyName ("review_date")]
-		public uint ReviewDate { get; set; }
+		public ulong ReviewDate { get; set; }
 
 		[JsonPropertyName ("rating")]
-		public uint Rating { get; set; }
+		public ulong Rating { get; set; }
 
 		[JsonPropertyName ("message")]
 		public string Message { get; set; }
